omve the left over parts to the getter/parser files
